Addresses issue 26 (fix config namespace) (#27)

+++ TagTracker.py
@@ -6,16 +6,10 @@
 import os
 import time
 import re
-from pathlib import Path
+import pathlib
 from typing import Tuple
-from TrackerConfig import (
-        all_tags,norm_tags,over_tags,retired_tags,
-        colour_letters,
-        CHECK_OUT_CONFIRM_TIME,T_UNDER,T_OVER,
-        CURSOR,INDENT,MODE_ROUND_TO_NEAREST,
-        statistics_kws,help_kws,audit_kws,edit_kws,del_kws,
-        query_kws,quit_kws,help_message,VERSION
-    )
+
+import trackerconfig
 
 def get_date() -> str:
     """Return current date as string: YYYY-MM-DD."""
@@ -44,25 +38,25 @@ def valid_time(inp:str) -> bool:
                 return True
     return False
 
-# FIXME - for later
 def canonical_tag( maybe_tag:str ) -> str|None:
     """Return 'maybe_tag' as canonical tag representation (or None).
 
     Canonical tag representation is a single colour letter, a tag letter
     and a sequence number, without lead zeroes.  All lowercase.
     """
+    # FIXME - for later
     if r := re.match(r"^ *([a-zA-z][a-zA-Z])0*([1-9][0-9]*) *$", maybe_tag ):
         return f"{r.group(1).lower()}{r.group(2)}"
     return None
 
-# FIXME - for later
 def valid_tag( tag:str ) ->bool:
     """Return whether 'tag' simplifies to a valid [known] tag."""
-    return bool(canonical_tag( tag ) in all_tags)
+    #FIXME: this is not used yet
+    return bool(canonical_tag( tag ) in trackerconfig.all_tags)
 
 def iprint(text:str, x:int=1) -> None:
     """Print the text, indented."""
-    print(f"{INDENT * x}{text}")
+    print(f"{trackerconfig.INDENT * x}{text}")
 
 def read_tags() -> bool:
     """Fetch tag data from file.
@@ -73,9 +67,8 @@ def read_tags() -> bool:
     if exists, read for ins and outs
     if none exists, make one.
     """
-
     #FIXME: Refactor
-    Path("logs").mkdir(exist_ok = True) # make logs folder if none exists
+    pathlib.Path("logs").mkdir(exist_ok = True) # make logs folder if none exists
     global check_ins, check_outs
     check_ins = {} # check in dictionary tag:time
     check_outs = {} # check out dictionary tag:time
@@ -88,7 +81,7 @@ def read_tags() -> bool:
             while line[0] != 'B': # while the first chr of each line isn't a header
                 cells = line.rstrip().split(',')
                 # if either a tag or time is invalid...
-                if not cells[0] in all_tags or not valid_time(cells[1]):
+                if not cells[0] in trackerconfig.all_tags or not valid_time(cells[1]):
                     print(f"Problem while reading {filedir} -- check-ins, line {line_counter}.")
                     return False
                 elif cells[0] in check_ins:
@@ -101,7 +94,7 @@ def read_tags() -> bool:
             line_counter += 1
             while line != '':
                 cells = line.rstrip().split(',')
-                if not cells[0] in all_tags or not valid_time(cells[1]):
+                if not cells[0] in trackerconfig.all_tags or not valid_time(cells[1]):
                     print(f"Problem while reading {filedir} -- check-outs, line {line_counter}.")
                     return False
                 elif cells[0] in check_outs:
@@ -120,7 +113,8 @@ def rotate_log() -> None:
     backuppath = f"{LOG_FILEPATH}.bak"
     if os.path.exists(backuppath):
         os.unlink(backuppath)
-    os.rename(LOG_FILEPATH,backuppath)
+    if os.path.exists(LOG_FILEPATH):
+        os.rename(LOG_FILEPATH,backuppath)
     return None
 
 def write_tags() -> None:
@@ -154,22 +148,23 @@ def minutes_to_time_str(time_in_minutes:int) -> str:
     time_as_text = f"{hours_portion}:{minutes_portion:02d}"
     return time_as_text
 
-#FIXME: this fn should now be unused & deprecated
+
 def minutes_to_time_str_list(times:list[int]) -> list[str]:
     """Convert list of times in minutes to list of HH:MM strings."""
     # FIXME: I think better to use minutes_to_time_str() than this
+    # FIXME: this fn should now be unused & deprecated
     text_times = []
     for time_in_minutes in times:
         text_times.append(minutes_to_time_str(time_in_minutes))
     return text_times
 
-#FIXME: this is not used yet
-def find_tag_durations(include_bikes_on_hand=True) -> dict[str:int]:
+def find_tag_durations(include_bikes_on_hand=True) -> dict[str,int]:
     """Make dict of tags with their stay duration in minutes.
 
-    include_bikes_on_hand, if True, will include checked in
+    If include_bikes_on_hand, this will include checked in
     but not returned out.  If False, only bikes returned out.
     """
+
     timenow = time_str_to_minutes(get_time())
     tag_durations = {}
     for tag,in_str in check_ins.items():
@@ -234,11 +229,11 @@ def mode_stay(stays:list) -> Tuple[int,int]:
     """
     round_stays = []
     for stay in stays:
-        remainder = stay % MODE_ROUND_TO_NEAREST
-        mult = stay // MODE_ROUND_TO_NEAREST
+        remainder = stay % trackerconfig.MODE_ROUND_TO_NEAREST
+        mult = stay // trackerconfig.MODE_ROUND_TO_NEAREST
         if remainder > 4:
             mult += 1
-        rounded = mult * MODE_ROUND_TO_NEAREST
+        rounded = mult * trackerconfig.MODE_ROUND_TO_NEAREST
         round_stays.append(rounded)
     mode = max(set(round_stays), key=round_stays.count)
     count = stays.count(mode)
@@ -249,9 +244,9 @@ def show_stats():
     norm = 0 # counting bikes by type
     over = 0
     for tag in check_ins:
-        if tag in norm_tags:
+        if tag in trackerconfig.norm_tags:
             norm += 1
-        elif tag in over_tags:
+        elif tag in trackerconfig.over_tags:
             over += 1
     tot_in = norm + over
 
@@ -275,14 +270,14 @@ def show_stats():
         medium_stays = 0
         long_stays = 0
         for stay in all_stays: # count stays over/under x time
-            if stay < T_UNDER:  # Changed from <= to < so matches description.
+            if stay < trackerconfig.T_UNDER:  # Changed from <= to < so matches description.
                 short_stays += 1
-            elif stay <= T_OVER:
+            elif stay <= trackerconfig.T_OVER:
                 medium_stays += 1
             else:
                 long_stays += 1
-        hrs_under = f"{(T_UNDER / 60):3.1f}" # times in hours for print clarity
-        hrs_over = f"{(T_OVER / 60):3.1f}"
+        hrs_under = f"{(trackerconfig.T_UNDER / 60):3.1f}" # times in hours for print clarity
+        hrs_over = f"{(trackerconfig.T_OVER / 60):3.1f}"
 
         iprint(f"\nSummary statistics as of {get_time()} "
                f"with {len(check_ins)-len(check_outs)} bikes still on hand:\n")
@@ -303,7 +298,7 @@ def show_stats():
         iprint(f"Mean stay:    {minutes_to_time_str(mean):>5}")
         iprint(f"Median stay:  {minutes_to_time_str(median):>5}")
         iprint(f"Mode stay:    {minutes_to_time_str(mode):>5} "
-               f"by {count} bike(s)  [{MODE_ROUND_TO_NEAREST} minute blocks]")
+               f"by {count} bike(s)  [{trackerconfig.MODE_ROUND_TO_NEAREST} minute blocks]")
 
     else: # don't try to calculate stats on nothing
         iprint("No bikes returned out, can't calculate statistics. "
@@ -313,13 +308,13 @@ def delete_entry(target = False, which_to_del=False, confirm = False) -> None:
     """Perform tag entry deletion dialogue."""
     del_syntax_message = ("Syntax: d <tag> <both or check-out only"
             " (b/o)> <optional pre-confirm (y)>")
-    if not(target in [False] + all_tags or which_to_del in [False, 'b', 'o']
+    if not(target in [False] + trackerconfig.all_tags or which_to_del in [False, 'b', 'o']
            or confirm in [False, 'y']):
         iprint(del_syntax_message) # remind of syntax if invalid input
         return None # interrupt
     if not target: # get target if unspecified
         iprint("Which tag's entry would you like to remove?")
-        target = input(f"(tag name) {CURSOR}").lower()
+        target = input(f"(tag name) {trackerconfig.CURSOR}").lower()
     checked_in = target in check_ins
     checked_out = target in check_outs
     if not checked_in and not checked_out:
@@ -332,14 +327,14 @@ def delete_entry(target = False, which_to_del=False, confirm = False) -> None:
                    f"a check-out ({time_out_temp}) recorded.")
             iprint("Do you want to delete (b)oth events, "
                    "or just the check-(o)ut?")
-            which_to_del = input(f"(b/o) {CURSOR}").lower()
+            which_to_del = input(f"(b/o) {trackerconfig.CURSOR}").lower()
         if which_to_del == 'b':
             if confirm == 'y': # pre-confirmation
                 sure = True
             else:
                 iprint("Are you sure you want to delete both events "
                        f"for {target}? (y/N)")
-                sure = input(f"(y/N) {CURSOR}").lower() == 'y'
+                sure = input(f"(y/N) {trackerconfig.CURSOR}").lower() == 'y'
             if sure:
                 check_ins.pop(target)
                 check_outs.pop(target)
@@ -354,7 +349,7 @@ def delete_entry(target = False, which_to_del=False, confirm = False) -> None:
             else:
                 iprint("Are you sure you want to delete the "
                        f"check-out record for {target}? (y/N)")
-                sure = input(f"(y/N) {CURSOR}").lower() == 'y'
+                sure = input(f"(y/N) {trackerconfig.CURSOR}").lower() == 'y'
 
             if sure:
                 time_temp = check_outs[target]
@@ -372,7 +367,7 @@ def delete_entry(target = False, which_to_del=False, confirm = False) -> None:
             else: # check
                 iprint("This tag has only a check-in recorded. "
                        "Are you sure you want to delete it? (y/N)")
-                sure = input(f"(y/N) {CURSOR}").lower() == 'y'
+                sure = input(f"(y/N) {trackerconfig.CURSOR}").lower() == 'y'
             if sure:
                 time_temp = check_ins[target]
                 check_ins.pop(target)
@@ -388,8 +383,8 @@ def query_tag(target = False, do_printing = True) -> str:
     #FIXME: what is do_printing meant to do?
     if not target: # only do dialog if no target passed
         iprint("Which tag would you like to query?")
-        target = input(f"(tag name) {CURSOR}").lower()
-    if target in retired_tags:
+        target = input(f"(tag name) {trackerconfig.CURSOR}").lower()
+    if target in trackerconfig.retired_tags:
         iprint(f"{target} has been retired.")
         return 'retired' # tag is retired
     else:
@@ -414,7 +409,7 @@ def prompt_for_time(inp = False) -> bool or str:
         iprint("What is the correct time for this event?")
         iprint("Use 24-hour format, or 'now' to use "
                f"the current time ({get_time()})")
-        inp = input(f"(HH:MM) {CURSOR}").lower()
+        inp = input(f"(HH:MM) {trackerconfig.CURSOR}").lower()
     if inp == 'now':
         return get_time()
     elif valid_time(inp):
@@ -431,16 +426,16 @@ def edit_entry(target = False, in_or_out = False, new_time = False):
             "<new time or 'now'>")
     if not target:
         iprint("Which bike's record do you want to edit?")
-        target = input(f"(tag ID) {CURSOR}").lower()
-    elif not target in all_tags:
+        target = input(f"(tag ID) {trackerconfig.CURSOR}").lower()
+    elif not target in trackerconfig.all_tags:
         iprint(edit_syntax_message)
         return False
-    if target in all_tags:
+    if target in trackerconfig.all_tags:
         if target in check_ins:
             if not in_or_out:
                 iprint("Do you want to change this bike's "
                        "check-(i)n or check-(o)ut time?")
-                in_or_out = input(f"(i/o) {CURSOR}").lower()
+                in_or_out = input(f"(i/o) {trackerconfig.CURSOR}").lower()
                 if not in_or_out in ['i','o']:
                     iprint(f"'{in_or_out}' needs to be 'i' or 'o' "
                            "(cancelled edit).")
@@ -490,22 +485,22 @@ def count_colours(inv:list[str]) -> str:
             if not char.isdigit():
                 shortened += char
         # cut off last, non-colour letter and add to list of abbrevs
-        if shortened in colour_letters:
+        if shortened in trackerconfig.colour_letters:
             just_colour_abbreviations.append(shortened)
         else:
             for x in range(10):
                 cutoff_by_x = shortened[:-x]
-                if cutoff_by_x in colour_letters:
+                if cutoff_by_x in trackerconfig.colour_letters:
                     just_colour_abbreviations.append(cutoff_by_x)
 
     colour_count = {} # build the count dictionary
-    for abbrev in colour_letters: # for each valid colour, loop through all tags
+    for abbrev in trackerconfig.colour_letters: # for each valid colour, loop through all tags
         if abbrev in just_colour_abbreviations:
             this_colour_count = 0
             for tag in just_colour_abbreviations:
                 if tag == abbrev:
                     this_colour_count += 1
-            colour_name = colour_letters[abbrev]
+            colour_name = trackerconfig.colour_letters[abbrev]
             colour_count[colour_name] = this_colour_count
 
     colour_count_str = '' # convert count dict to string
@@ -552,7 +547,7 @@ def tag_check(tag:str) -> None:
 
     This processes a prompt that's just a tag ID.
     """
-    if not tag in retired_tags:
+    if not tag in trackerconfig.retired_tags:
         checked_in = tag in check_ins
         checked_out = tag in check_outs
         if checked_in:
@@ -560,7 +555,7 @@ def tag_check(tag:str) -> None:
                 query_tag(tag)
                 iprint(f"Overwrite {check_outs[tag]} check-out with "
                        f"current time ({get_time()})? (y/N)")
-                sure = input(f"(y/N) {CURSOR}") == 'y'
+                sure = input(f"(y/N) {trackerconfig.CURSOR}") == 'y'
                 if sure:
                     edit_entry(target = tag, in_or_out = 'o',
                             new_time = get_time())
@@ -570,11 +565,11 @@ def tag_check(tag:str) -> None:
                 now_mins = time_str_to_minutes(get_time())
                 check_in_mins = time_str_to_minutes(check_ins[tag])
                 time_diff_mins = now_mins - check_in_mins
-                if time_diff_mins < CHECK_OUT_CONFIRM_TIME: # if < 1/2 hr
+                if time_diff_mins < trackerconfig.CHECK_OUT_CONFIRM_TIME: # if < 1/2 hr
                     iprint("This bike checked in at "
                            f"{check_ins[tag]} ({time_diff_mins} mins ago).")
                     iprint("Do you want to check it out? (Y/n)")
-                    sure = input(f"(Y/n) {CURSOR}").lower() in ['', 'y']
+                    sure = input(f"(Y/n) {trackerconfig.CURSOR}").lower() in ['', 'y']
                 else: # don't check for long stays
                     sure = True
                 if sure:
@@ -598,13 +593,13 @@ def process_prompt(prompt:str) -> None:
         kwd = cells[0] # take first phrase as fn to call
         if cells[0][0] in ['?', '/'] and len(cells[0])>1: # take non-letter query prompts without space
             query_tag(cells[0][1:], False)
-        elif kwd in statistics_kws:
+        elif kwd in trackerconfig.statistics_kws:
             show_stats()
-        elif kwd in help_kws:
-            print(help_message)
-        elif kwd in audit_kws:
+        elif kwd in trackerconfig.help_kws:
+            print(trackerconfig.help_message)
+        elif kwd in trackerconfig.audit_kws:
             show_audit()
-        elif kwd in edit_kws:
+        elif kwd in trackerconfig.edit_kws:
             args = len(cells) - 1 # number of arguments passed
             target, in_or_out, new_time = None, None, None# initialize all
             if args > 0:
@@ -616,7 +611,7 @@ def process_prompt(prompt:str) -> None:
             edit_entry(target = target, in_or_out = in_or_out,
                     new_time = new_time)
 
-        elif kwd in del_kws:
+        elif kwd in trackerconfig.del_kws:
             args = len(cells) - 1 # number of arguments passed
             target, which_to_del, pre_confirm = False, False, False
             if args > 0:
@@ -628,14 +623,14 @@ def process_prompt(prompt:str) -> None:
             delete_entry(target = target, which_to_del = which_to_del,
                     confirm = pre_confirm)
 
-        elif kwd in query_kws:
+        elif kwd in trackerconfig.query_kws:
             try:
                 query_tag(target = cells[1]) # query the tag that follows cmd
             except IndexError:
                 query_tag() # if no tag passed run the dialog
-        elif kwd in quit_kws:
+        elif kwd in trackerconfig.quit_kws:
             exit() # quit program
-        elif kwd in all_tags:
+        elif kwd in trackerconfig.all_tags:
             tag_check(kwd)
         else: # not anything recognized so...
             iprint(f"'{prompt}' isn't a recognized tag or command "
@@ -648,7 +643,7 @@ def main() -> None:
     rotate_log()
     write_tags() # save before input regardless
     #show_audit() # show all bikes currently in
-    prompt = input(f"\n\nEnter a tag or command {CURSOR}").lower() # take input
+    prompt = input(f"\n\nEnter a tag or command {trackerconfig.CURSOR}").lower() # take input
     process_prompt(prompt)
     main() # loop
 
@@ -656,12 +651,12 @@ def main() -> None:
 check_ins = {}
 check_outs = {}
 
-print(f"TagTracker {VERSION} by Julias Hocking")
+print(f"TagTracker {trackerconfig.VERSION} by Julias Hocking")
 DATE = get_date()
-LOG_FILEPATH = f"logs/{LOG_BASENAME}{DATE}.log"
+LOG_FILEPATH = f"logs/{trackerconfig.LOG_BASENAME}{DATE}.log"
 
 if read_tags(): # only run main() if tags read successfully
     main()
 else: # if read_tags() problem
-    print(f"\n{INDENT}Closing automatically in 30 seconds...")
+    print(f"\n{trackerconfig.INDENT}Closing automatically in 30 seconds...")
     time.sleep(30)
